backend/requirements.txt
Flask==2.0.1
mysql-connector-python==8.1.0
Werkzeug==2.2.2